Add external openshift client
Created DeployedOpenshiftClient for accessing the caching-service within
an application deployed on an openshift pod.

Created ExternalOpenshiftClient for accessing the caching-service on
machines external to the openshift cluster running the caching-service.
The creation of the local truststore is achieved by utilising the oc client
to retrieve the services certificates from the remote openshift cluster.